elimine menu.js
