--- Menu.js
@@ -1,8 +1,0 @@
-const menuHamburguesas = [
-    { id: 1, nombre: "Hamburguesa Clásica", precio: 8500 },
-    { id: 2, nombre: "Hamburguesa Texana", precio: 9500 },
-    { id: 3, nombre: "Hamburguesa de Queso", precio: 10000 },
-    { id: 4, nombre: "Hamburguesa Vegana", precio: 12000 },
-    ];
-
-    export default menuHamburguesas;
